+++ src/utils/createRequestHeaders.js
@@ -1,17 +1,5 @@
-export default async ({ getAccessToken, version = undefined } = {}) => {
-  const accessToken = await getAccessToken();
-
-  const headers = {
-    'Content-Type': 'application/json',
-    Authorization: `Bearer ${accessToken}`,
-  };
-
-  if (version !== undefined) {
-    return {
-      ...headers,
-      'X-Api-Version': version,
-    };
-  }
-
-  return headers;
-};
+export default async ({ getAccessToken, version } = {}) => ({
+  'Content-Type': 'application/json',
+  ...(getAccessToken && { Authorization: `Bearer ${await getAccessToken()}` }),
+  ...(version && { 'X-Api-Version': version }),
+});